#45 - Add an encoder to test <form> data delivery
FORM DATA must be serialized and used in a different form than the API.
Create a test-only utility that converts objects to form data-formatted strings to conform to the conventions. utility dedicated to testing, and write tests for it.
In the absence of this, you can put in the form `post().param()` values, but that doesn't seem to match the actual post request transmission. Also, it's inconvenient because you have to put it as a parameter, not an object.