#62 - Improving AuthControllerTest: Void.class
This is supposed to be a bean, not a class.
Since `Void` is not a Spring bean, I realized that this is not a correct representation and causes errors in some development environments.
I improved this by creating a proper test-only bean.
Test beans don't have content inside them because their purpose is to do nothing.
Considering that it is unlikely to be recycled by other tests, it is written as an inner class.