#68 - Refactoring: Change entity association field wrapping
I chose to write a single line because writing in multiple lines would result in longer lines of code,
As the requirements grow and more relationships are added.
the annotations and code patterns become increasingly different.
The alignment doesn't look good

So I changed only the relation field to a plain multi-line style first.