#69 - [Advanced]Apply getters to entity equals() and hashcode()
If you access the field directly here,
When dealing with proxy objects created by Hibernate for lazy loading
Field values may be `null`.
Then we can't do the comparison logic properly, so we need to use the getter.
This ensures that we can read the value correctly even if we're dealing with a proxy object.