#74 - Implementing the Hashtag Service
Implemented functionality for managing hashtags attached to posts.
Uniquely, there is no method to handle saving hashtags.
Saving a hashtag is a function that only happens when a post is saved or modified.
It only happens when a post is saved or modified.
This is all handled by the post service via association mapping.

* Search DB for hashtag information by hashtag name
* Parsing hashtags from the body
* Delete a hashtag that no longer exists in a post